revert ForceReply
