+++ autobot/telegram/objects/__init__.py
@@ -36,6 +36,7 @@ from .contact import Contact
 from .dice import Dice
 from .document import Document
 from .file import File
+from .forcereply import ForceReply
 from .forumtopic import ForumTopicClosed, ForumTopicCreated, ForumTopicReopened
 from .inlinekeyboard import InlineKeyboardButton, InlineKeyboardMarkup
 from .inputfile import (
@@ -60,7 +61,7 @@ from .message import Message, MessageAutoDeleteTimerChanged, MessageEntity, Mess
 from .photosize import PhotoSize
 from .poll import Poll, PollAnswer, PollOption
 from .proximityalerttriggered import ProximityAlertTriggered
-from .replykeyboard import ForceReply, ReplyKeyboardMarkup, ReplyKeyboardRemove
+from .replykeyboard import ReplyKeyboardMarkup, ReplyKeyboardRemove
 from .responseparameters import ResponseParameters
 from .update import Update
 from .user import User, UserProfilePhotos
@@ -77,81 +78,82 @@ from .voice import Voice
 from .webapp import WebAppData, WebAppInfo
 
 __all__ = ("Animation",
-           "Audio",
-           "BaseObject",
-           "BotCommand",
-           "BotCommandScope",
-           "BotCommandScopeAllChatAdministrators",
-           "BotCommandScopeAllGroupChats",
-           "BotCommandScopeAllPrivateChats",
-           "BotCommandScopeChat",
-           "BotCommandScopeChatAdministrators",
-           "BotCommandScopeChatMember",
-           "BotCommandScopeDefault",
-           "CallBackQuery",
-           "Chat",
-           "ChatAdministratorRights",
-           "ChatInviteLink",
-           "ChatJoinRequest",
-           "ChatLocation",
-           "ChatPermissions",
-           "ChatPhoto",
-           "ChatMember",
-           "ChatMemberAdministrator",
-           "ChatMemberBanned",
-           "ChatMemberLeft",
-           "ChatMemberMember",
-           "ChatMemberOwner",
-           "ChatMemberRestricted",
-           "ChatMemberUpdated",
-           "Contact",
-           "Dice",
-           "Document",
-           "File",
-           "ForceReply",
-           "ForumTopicClosed",
-           "ForumTopicCreated",
-           "ForumTopicReopened",
-           "InlineKeyboardButton",
-           "InlineKeyboardMarkup",
-           "InputFile",
-           "InputMedia",
-           "InputMediaAnimation",
-           "InputMediaAudio",
-           "InputMediaDocument",
-           "InputMediaPhoto",
-           "InputMediaVideo",
-           "KeyboardButton",
-           "KeyboardButtonPollType",
-           "Location",
-           "LoginUrl",
-           "MenuButton",
-           "MenuButtonCommands",
-           "MenuButtonDefault",
-           "MenuButtonWebApp",
-           "Message",
-           "MessageAutoDeleteTimerChanged",
-           "MessageEntity",
-           "MessageId",
-           "PhotoSize",
-           "Poll",
-           "PollAnswer",
-           "PollOption",
-           "ProximityAlertTriggered",
-           "ReplyKeyboardMarkup",
-           "ReplyKeyboardRemove",
-           "ResponseParameters",
-           "Update",
-           "User",
-           "UserProfilePhotos",
-           "Venue",
-           "Video",
-           "VideoChatEnded",
-           "VideoChatParticipantsInvited",
-           "VideoChatScheduled",
-           "VideoChatStarted",
-           "VideoNote",
-           "Voice",
-           "WebAppData",
-           "WebAppInfo",
-           )
+            "Audio",
+            "BaseObject",
+            "BotCommand",
+            "BotCommandScope",
+            "BotCommandScopeAllChatAdministrators",
+            "BotCommandScopeAllGroupChats",
+            "BotCommandScopeAllPrivateChats",
+            "BotCommandScopeChat",
+            "BotCommandScopeChatAdministrators",
+            "BotCommandScopeChatMember",
+            "BotCommandScopeDefault",
+            "CallBackQuery",
+            "Chat",
+            "ChatAdministratorRights",
+            "ChatInviteLink",
+            "ChatJoinRequest",
+            "ChatLocation",
+            "ChatPermissions",
+            "ChatPhoto",
+            "ChatMember",
+            "ChatMemberAdministrator",
+            "ChatMemberBanned",
+            "ChatMemberLeft",
+            "ChatMemberMember",
+            "ChatMemberOwner",
+            "ChatMemberRestricted",
+            "ChatMemberUpdated",
+            "Contact",
+            "Dice",
+            "Document",
+            "File",
+            "ForceReply",
+            "ForumTopicClosed",
+            "ForumTopicCreated",
+            "ForumTopicReopened",
+            "InlineKeyboardButton",
+            "InlineKeyboardMarkup",
+            "InputFile",
+            "InputMedia",
+            "InputMediaAnimation",
+            "InputMediaAudio",
+            "InputMediaDocument",
+            "InputMediaPhoto",
+            "InputMediaVideo",
+            "KeyboardButton",
+            "KeyboardButtonPollType",
+            "Location",
+            "LoginUrl",
+            "MenuButton",
+            "MenuButtonCommands",
+            "MenuButtonDefault",
+            "MenuButtonWebApp",
+            "Message",
+            "MessageAutoDeleteTimerChanged",
+            "MessageEntity",
+            "MessageId",
+            "PhotoSize",
+            "Poll",
+            "PollAnswer",
+            "PollOption",
+            "ProximityAlertTriggered",
+            "ReplyKeyboardMarkup",
+            "ReplyKeyboardRemove",
+            "ResponseParameters",
+            "Update",
+            "User",
+            "UserProfilePhotos",
+            "Venue",
+            "Video",
+            "VideoChatEnded",
+            "VideoChatParticipantsInvited",
+            "VideoChatScheduled",
+            "VideoChatStarted",
+            "VideoNote",
+            "Voice",
+            "WebAppData",
+            "WebAppInfo",
+        )
+
